minor Ui, starting signup

+++ lib/student details/student_details.dart
@@ -1,4 +1,5 @@
-import 'package:classmyte/data_management/data_retrieval.dart';
+import 'dart:io';
+
 import 'package:classmyte/data_management/edit_contacts.dart';
 import 'package:flutter/material.dart';
 
@@ -12,7 +13,6 @@ class StudentDetailsScreen extends StatefulWidget {
 }
 
 class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
-  // ValueNotifiers for student data
   late final ValueNotifier<String> nameNotifier;
   late final ValueNotifier<String> fatherNameNotifier;
   late final ValueNotifier<String> classNotifier;
@@ -37,9 +37,37 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
     getStudentData();
   }
 
- void deleteStudent() {
-    EditContactService.deleteContact(widget.student['id'] ?? '');
-    Navigator.pop(context);
+  void deleteStudent() {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Confirm Delete'),
+          content: const Text('Are you sure you want to delete this contact?'),
+          actions: <Widget>[
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(false);
+              },
+              child: const Text('No'),
+            ),
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(true);
+              },
+              child: const Text('Yes'),
+            ),
+          ],
+        );
+      },
+    ).then((confirmed) {
+      if (confirmed == true) {
+        EditContactService.deleteContact(widget.student['id'] ?? '');
+        Navigator.pop(context);
+      } else {
+        // print('Deletion canceled');
+      }
+    });
   }
 
   void saveChanges() {
@@ -54,8 +82,8 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
       altNumberNotifier.value,
     );
     isEditableNotifier.value = false;
-    getStudentData(); 
-    Navigator.pop(context); 
+    getStudentData();
+    Navigator.pop(context);
   }
 
   Future<void> getStudentData() async {
@@ -63,7 +91,8 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
     isLoadingNotifier.value = false;
   }
 
-  void _selectDate(BuildContext context, ValueNotifier<String> dateNotifier) async {
+  void _selectDate(
+      BuildContext context, ValueNotifier<String> dateNotifier) async {
     DateTime? picked = await showDatePicker(
       context: context,
       initialDate: DateTime.now(),
@@ -79,17 +108,24 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.student['name']!),
+        backgroundColor: Colors.blue,
+        title: const Text("Edit Details",  style: TextStyle(fontSize: 26, fontWeight: FontWeight.bold, color: Colors.white)),
         actions: [
-          IconButton(
-            icon: const Icon(Icons.edit),
-            onPressed: () {
-              isEditableNotifier.value = true; // Enable editing
+          ValueListenableBuilder<bool>(
+            valueListenable: isEditableNotifier,
+            builder: (context, isEditable, child) {
+              return isEditable
+                  ? IconButton(
+                      icon: const Icon(Icons.save, color: Colors.white,),
+                      onPressed: () => saveChanges(),
+                    )
+                  : IconButton(
+                      icon: const Icon(Icons.edit,  color: Colors.white,),
+                      onPressed: () {
+                        isEditableNotifier.value = true; // Enable editing
+                      },
+                    );
             },
-          ),
-          IconButton(
-            icon: const Icon(Icons.save),
-            onPressed: () => saveChanges(),
           ),
         ],
       ),
@@ -102,35 +138,63 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
             return Padding(
               padding: const EdgeInsets.all(16.0),
               child: Column(
+                
                 children: [
-                  CircleAvatar(
-                    radius: 50,
-                    backgroundImage: widget.student['profilePictureUrl'] != null
-                        ? NetworkImage(widget.student['profilePictureUrl']!)
-                        : null,
-                    child: widget.student['profilePictureUrl'] == null
-                        ? const Icon(Icons.person, size: 50)
-                        : null,
+                  
+                  Stack(
+                    
+                    alignment: Alignment.center,
+                    children: [
+                      CircleAvatar(
+                        radius: 30,
+                        backgroundColor: Colors.blue,
+
+                            child:  Text(
+                                nameNotifier.value[0],
+                                style: const TextStyle(
+                                    color: Colors.white, fontSize: 24),
+                              )
+                      ),
+                    ],
                   ),
                   const SizedBox(height: 16),
                   Expanded(
-                    child: ListView(
-                      children: [
-                        _buildEditableTile('Father Name', fatherNameNotifier),
-                        _buildEditableTile('Class', classNotifier),
-                        _buildEditableTile('Phone#', phoneNumberNotifier, isPhone: true),
-                        _buildEditableTile('Alternate#', altNumberNotifier, isPhone: true),
-                        _buildDateTile('Date of Birth', dobNotifier),
-                        _buildDateTile('Admission Date', admissionNotifier),
-                      ],
+                    
+                    child: Center(
+                      child: ListView(
+                        
+                        children: [
+                          Container(
+                              decoration: BoxDecoration(
+                                  border: Border.all(color: Colors.black)),
+                              child: Column(
+
+                                children: [
+                                  _buildEditableTile('Name', nameNotifier),
+                                  _buildEditableTile(
+                                      'Father Name', fatherNameNotifier),
+                                  _buildEditableTile('Class', classNotifier),
+                                  _buildEditableTile(
+                                      'Phone#', phoneNumberNotifier,
+                                      isPhone: true),
+                                  _buildEditableTile(
+                                      'Alternate#', altNumberNotifier,
+                                      isPhone: true),
+                                  _buildDateTile('Date of Birth', dobNotifier),
+                                  _buildDateTile(
+                                      'Admission Date', admissionNotifier),
+                                ],
+                              )),
+                        ],
+                      ),
                     ),
                   ),
                   ElevatedButton(
                     onPressed: deleteStudent,
                     style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.red,
+                      backgroundColor: Colors.blue,
                     ),
-                    child: const Text('Delete'),
+                    child: const Icon(Icons.delete, color: Colors.white,),
                   ),
                 ],
               ),
@@ -141,28 +205,36 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
     );
   }
 
+  Widget _buildEditableTile(String title, ValueNotifier<String> valueNotifier,
+      {bool isPhone = false}) {
+    TextEditingController controller = TextEditingController(
+        text: valueNotifier
+            .value); // Initialize controller with the current value
 
-  Widget _buildEditableTile(String title, ValueNotifier<String> valueNotifier, {bool isPhone = false}) {
     return ValueListenableBuilder<bool>(
       valueListenable: isEditableNotifier,
       builder: (context, isEditable, child) {
         return ValueListenableBuilder<String>(
           valueListenable: valueNotifier,
           builder: (context, value, child) {
+            if (isEditable) {
+              controller.value = TextEditingValue(text: value);
+            }
             return ListTile(
-              title: Text(title),
+              title:Text(title, style: const TextStyle(fontWeight: FontWeight.bold),),
               subtitle: isEditable
                   ? TextField(
+                      controller: controller,
                       onChanged: (newValue) {
                         valueNotifier.value = newValue;
                       },
-                      decoration: InputDecoration(
-                        border: const OutlineInputBorder(),
-                        hintText: 'Enter $title',
+                      decoration: const InputDecoration(
+                        border: OutlineInputBorder(),
                       ),
-                      keyboardType: isPhone ? TextInputType.phone : TextInputType.text,
+                      keyboardType:
+                          isPhone ? TextInputType.phone : TextInputType.text,
                     )
-                  : Text(value, style: const TextStyle(fontSize: 16)),
+                  : Text(value, style: const TextStyle(fontSize: 16), ),
             );
           },
         );
@@ -178,9 +250,11 @@ class _StudentDetailsScreenState extends State<StudentDetailsScreen> {
           valueListenable: dateNotifier,
           builder: (context, value, child) {
             return ListTile(
-              title: Text(title),
+              title: Text(title,  style: const TextStyle(fontWeight: FontWeight.bold),),
               subtitle: GestureDetector(
-                onTap: isEditable ? () => _selectDate(context, dateNotifier) : null,
+                onTap: isEditable
+                    ? () => _selectDate(context, dateNotifier)
+                    : null,
                 child: Text(
                   value.isNotEmpty ? value : 'Select $title',
                   style: const TextStyle(fontSize: 16),
